split encoder  to separate file
