+++ users/jonavin/readme.md
@@ -35,6 +35,7 @@ KEYCODES:
 - KC_SFTUP        RShift when held, Up arrow when tapped
 - KC_RAISESPC     _RAISE layer mod when held, space when tapped
 - KC_LOWERSPC     _LOWER layer mod when held, space when tapped
+- KC_TSTOG        toggles between volume and Alt-Tab encoder control
 - 
  When EMOTICON_ENABLE = yes
 -   EMO_SHRUG       `\_("/)_/`
@@ -61,6 +62,7 @@ AVAILABLE ENCODER ACTIONS:
 - void encoder_action_rgb_mode(bool clockwise);
 -
 - void encoder_action_alttabscroll(bool clockwise)
+- void encoder_toggle_alttabscroll(void);
 
 ENABLE FEATURES your keymap rules.mk
 ---------------------------------------
